Foreign Direct Investment
The project has a major goal to investigate various factors that are connected to the labor force and the economic freedom index, as well as the gross domestic product (GDP), in order to determine their impact on the inflow of foreign direct investment (FDI) into the country.

Foreign direct investment refers to the investment made by foreign entities, such as individuals or corporations, into a country's economy. These investments may be made in various forms, such as building factories or acquiring ownership of a local business.

The project aims to examine how different factors, such as labor force availability and skills, economic freedom index, and GDP, may influence the inflow of FDI into the country. This investigation can help to identify the specific aspects of the country's economy that may be attractive or unattractive to foreign investors, and thereby inform policy decisions aimed at increasing FDI inflows.

The potential benefits of increased FDI inflows include increased employment opportunities, increased economic growth, and improved technological development in the country. Therefore, the findings of this study can be valuable in guiding policy decisions aimed at improving the country's economic performance and overall well-being.